requirements.txt
boto3>=1.26.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pika>=1.3.0
databricks-sql-connector>=2.0.0
databricks-sdk>=0.12.0
pandas>=1.5.0
